+++ R/aml_collect_data.R
@@ -38,7 +38,7 @@
 #' path_data <- normalizePath(tempdir(),winslash = "/")
 #'
 #'
-#' # performing Deep Learning Regression using the custom function
+#' # data transformation using the custom function for one symbol
 #' aml_collect_data(price_dataset = prices,
 #'                  indicator_dataset = macd,
 #'                  symbol = 'EURUSD',